// emp-mgm/src/app/model/stockinstruments.model.ts
export class StockInstruments {
    _id?: string;
    exchange?: string;
    instrument_type?: string;
    name?: string;
    stock_name?: string;
    segment?: string;
    last_price?: number;
    tradingsymbol?: string;
    last_synced?: Date;
    fundamentalAnalysis?: string;
    technicalAnalysis?: string;
    ISIN?: string;
    market_cap?: string;
    sector_name?: string;
    invesmateAnalysis?: string;
}